requirements.txt
setuptools
wheel
aiohttp
discord
python-dotenv
